libbcc: replacing sqrt() calls  with faster vsqrt instruction
On LLVM compiler, only unmangled sqrt()/sqrtf() will be recognized as
an intrinsic, which will eventually be emitted as an "vsqrt" instr if
possible.  However, all function names  in user's code are manged.
In this patch, we provide a sqrt(), which calls sqrtf() that is
declared as extern "C". So, all calls to sqrt() in user's code, which
are originally routed to libm, will be replaced as sqrtf(), which will
furthur be replaced by vsqrt instruction.

Change-Id: I1cd4690305c531f07e61d02c3f0c7fb3c1445578